Ticket#13632,CP-2393,Ajuste bloqueo de monto a operar

+++ custom/modules/Opportunities/clients/base/views/create/create.js
@@ -2424,7 +2424,7 @@
         //VALIDA QUE NO SUPERE EL $1,000,000 (UN MILLON) EN EL CAMPO DEL MONTO DEL TIPO DE PRODUCTO TARJETA DE CREDITO - CREATE
         if (this.model.get('tipo_producto_c') == '14') {
 
-            if (parseInt(this.model.get('monto_c')) > parseInt(this.model.get('control_monto_c'))) {
+            if (parseFloat(this.model.get('monto_c')) > parseFloat(this.model.get('control_monto_c'))) {
 
                 app.alert.show('message-control-monto', {
                     level: 'error',
